Fixed naming convention for dns and removed start command since we wont use it
